refactor(in_app_webview): extract legacy dismissDialogues into testable seam
Move the inline onLoadStop JS and retry loop into dismiss_dialogues_legacy.dart
(legacyDismissDialoguesJs / shouldDismissDialogues / runLegacyDismissDialogues) so
the legacy behavior can be characterized. Behavior is unchanged.
